Address QE and staff review feedback on tile cache feature

+++ composeApp/src/commonMain/kotlin/com/jordankurtz/piawaremobile/map/debug/TileCacheStats.kt
@@ -1,13 +1,13 @@
 package com.jordankurtz.piawaremobile.map.debug
 
 data class TileCacheStats(
-    val diskHits: Int = 0,
-    val networkFetches: Int = 0,
-    val errors: Int = 0,
+    val diskHits: Long = 0L,
+    val networkFetches: Long = 0L,
+    val errors: Long = 0L,
 ) {
     /** Total resolved tile requests (hits + misses, excluding errors). */
-    val total: Int get() = diskHits + networkFetches
+    val total: Long get() = diskHits + networkFetches
 
     /** Fraction of tile requests served from disk cache (0.0–1.0). */
-    val hitRate: Float get() = if (total == 0) 0f else diskHits.toFloat() / total
+    val hitRate: Float get() = if (total == 0L) 0f else diskHits.toFloat() / total
 }
